requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
sqlalchemy==2.0.34
pydantic==2.8.2
python-dotenv==1.0.1
